Python implementation working

+++ src/bindings/pyagent.rs
@@ -11,7 +11,7 @@ pub struct PyAgent {
 #[pymethods]
 impl PyAgent {
     #[getter]
-    fn num(&self) -> u32 {
+    fn num(&self) -> usize {
         self.agent.id()
     }
